--- chembl2uniprot.egg-info/PKG-INFO
@@ -1,18 +1,0 @@
-Metadata-Version: 2.4
-Name: chembl2uniprot
-Version: 0.1.0
-Summary: Map ChEMBL IDs to UniProt IDs
-Author-email: AutoGen <noreply@example.com>
-Requires-Python: >=3.10
-Requires-Dist: pandas
-Requires-Dist: requests
-Requires-Dist: PyYAML
-Requires-Dist: jsonschema
-Requires-Dist: tqdm
-Requires-Dist: tenacity
-Provides-Extra: dev
-Requires-Dist: pytest; extra == "dev"
-Requires-Dist: requests-mock; extra == "dev"
-Requires-Dist: black; extra == "dev"
-Requires-Dist: ruff; extra == "dev"
-Requires-Dist: mypy; extra == "dev"
