--- chembl2uniprot.egg-info/SOURCES.txt
@@ -1,12 +1,0 @@
-README.md
-pyproject.toml
-chembl2uniprot/__init__.py
-chembl2uniprot/__main__.py
-chembl2uniprot/config.py
-chembl2uniprot/mapping.py
-chembl2uniprot.egg-info/PKG-INFO
-chembl2uniprot.egg-info/SOURCES.txt
-chembl2uniprot.egg-info/dependency_links.txt
-chembl2uniprot.egg-info/requires.txt
-chembl2uniprot.egg-info/top_level.txt
-tests/test_mapping.py
