--- chembl2uniprot.egg-info/dependency_links.txt
@@ -1,1 +1,0 @@
-
