--- chembl2uniprot.egg-info/requires.txt
@@ -1,13 +1,0 @@
-pandas
-requests
-PyYAML
-jsonschema
-tqdm
-tenacity
-
-[dev]
-pytest
-requests-mock
-black
-ruff
-mypy
