--- chembl2uniprot.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-chembl2uniprot
