+++ library/chembl2uniprot/config.py
