+++ library/chembl2uniprot/mapping.py
