--- mypy.ini
@@ -1,2 +1,0 @@
-[mypy]
-ignore_missing_imports = True
