+++ pyproject.toml
@@ -26,5 +26,8 @@ dev = [
     "mypy",
 ]
 
-[tool.setuptools.package-data]
-"chembl2uniprot" = ["default_config.yaml", "config.schema.json"]
+[tool.setuptools.packages.find]
+where = ["library"]
+
+[tool.mypy]
+ignore_missing_imports = true
